+++ resources/classes/Database/DatabaseException.class.php
@@ -1,0 +1,8 @@
+<?php
+namespace Database;
+
+	class DatabaseException extends \Exception {
+		public function __construct($message, $code = 0, Exception $previous = null) {
+			parent::__construct($message, $code, $previous);
+		}
+	}